academic-chatbot/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.11.0
pandas==2.3.3
numpy==2.3.3
pypdf2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.50.0
google-generativeai==0.8.3
python-dotenv==1.0.0
matplotlib==3.10.6
seaborn==0.12.2
huggingface-hub==0.35.3